+++ noteconsole/noteconsole/plugins.cs
@@ -1,0 +1,10 @@
+namespace noteconsole
+{
+    public interface IPlugin
+    {
+        PluginInfo GetPluginInfo();
+        List<ColorsGlobal> MainFunction();
+    }
+
+
+}